requirements.txt
# Path: codeSage.be/requirements.txt

# Core Dependencies
Flask==3.0.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Sock==0.7.0
python-dotenv==1.0.0
cryptography==42.0.0
SQLAlchemy==2.0.25
aiosqlite==0.19.0

# AI and ML Dependencies
torch>=2.2.0
transformers==4.36.2
tokenizers==0.15.1
numpy==1.26.3
scipy==1.12.0

# Code Analysis
gitpython==3.1.41
chardet==5.2.0
pylint==3.0.3

# Visualization and Data Processing
networkx==3.2.1
plotly==5.18.0
pandas==2.2.0
matplotlib==3.8.2

# WebSocket and Async
websockets==12.0
simple-websocket==1.0.0
aiohttp==3.9.3

# Performance and Monitoring
psutil==5.9.8

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0